Creacion de Gestión de Objetivos de Aprendizaje con sus carpetas y archivos correspondiente

+++ micro_academico/src/main/java/com/example/micro_academico/gestionAsignatura/model/request/RegistrarAsignaturaRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class RegistrarAsignaturaRequest {
-    private Long idAsignatura;
+    private int idAsignatura;
     private String nombreAsignatura;
     private Integer horasSemanales;
     private String niverRequerido;
